requirements.txt
click
platformdirs
six
psutil
wandb
ruamel.yaml
configmypy
tensorly
tensorly-torch
torch-harmonics
matplotlib
opt-einsum
h5py
zarr
